Reorganized raw socket functions into a single file.
